Add browser support guide
The FEWD team has spent the last few months developing some rules and
guidance around browser support. This guide aims to document and share
those with the greater dev team as well as act as a reference for future
team members. It is a work in progress and will be updated as we
continue exploring this topic.