Add documentation for tracking changes to Django and Wagtail
In this change I have attempted to document the process we've undertaken to upgrade Django and Wagtail.

I've moved the supported-versions matrix from the Python standard doc and makes it more concrete to apply to Django and Wagtail and cf.gov and its satellites and libraries.

I've documented the approach we have been taking to feature detection and version testing, and the actual approach to keeping satellites, libraries, and cf.gov itself up-to-date.

It's a little bit ARID with the changes to the Python standards doc and tox, but that's intentional.

Co-authored-by: Scott Cranfill <scott.cranfill@cfpb.gov>